requirements.txt
selenium == 4.29.0
pytest == 8.3.5
allure-pytest == 2.13.5
